requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
sqlalchemy>=2.0
alembic>=1.13
pydantic-settings>=2.0
pydantic[email]>=2.0
python-jose[cryptography]>=3.3
bcrypt>=4.0
python-multipart>=0.0.9

# Dev / testing
pytest>=8.0
httpx>=0.27
